requirements.txt
torch==1.8.0
torchvision==0.9.0

Flask==2.0.2
matplotlib==3.4.3
numpy
onnx==1.10.2
onnx-tf==1.9.0
onnxruntime-gpu==1.10.0
opencv-python==4.5.3.56
Pillow==8.4.0
requests==2.26.0
timm==0.4.12
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
requests
urllib3==1.26.8